- 히스토리 api 추가 - 운동에 대한 카테고리(~= 키워드) 추가 - jwt 토큰방식 추가로 인한 기본 루틴, 운동, 히스토리 로직 변경 - 루틴 정책 변경(등록, 기록 모두 가능)

+++ src/main/java/com/ms/wms/domain/routine_exercise/RoutineExercise.java
@@ -30,12 +30,12 @@ public class RoutineExercise {
     private int count;
 
     @Column
-    private int weight;
+    private Double weight;
 
     @Column
     private int exerciseSet;
 
-    public static RoutineExercise createSaveRoutineExercise(Routine routine, Exercise exercise, int exerciseSet, int count, int weight) {
+    public static RoutineExercise createSaveRoutineExercise(Routine routine, Exercise exercise, int exerciseSet, int count, Double weight) {
         RoutineExercise routineExercise = new RoutineExercise();
         routineExercise.routine = routine;
         routineExercise.exercise = exercise;
